// web/src/app/core/api/generated/models/api-validation-problem-details.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ApiValidationProblemDetails {
  code?: string;
  detail?: (string | null);
  errors?: {
[key: string]: Array<string>;
};
  instance?: (string | null);
  status?: (number | string | null);
  title?: (string | null);
  traceId?: string;
  type?: (string | null);
}